pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-log-triage-env"
version = "0.3.0"
description = "Log Analysis and Incident Triage Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-3-Clause" }
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
server = "mastermind_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["mastermind_env", "mastermind_env.server"]
package-dir = { "mastermind_env" = ".", "mastermind_env.server" = "server" }
